summar for question 32, 33 and 34
Summarized each question but did not remove dates after dec 31st. Wrote code that removes dates after 31st but just have to update the quarto file in order to render a word doc with the correct output. n should not equal 488 for any of the quetions, if it does then this means you did not delete dates after dec 31st (2022)